requirements.txt
streamlit
langchain_community
openai
python-dotenv
